# API/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoffeeFlow API Docs",
    "description": "Documentacion interactiva completa de la API de CoffeeFlow: Autenticacion, Usuarios, Mesas, Productos, Pedidos, Cocina (ingredientes/recetas/inventario/alertas), Caja (turnos/tickets/pagos/cortes/gastos/compras) y Notificaciones.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Servidor Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/pedidos": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Listar pedidos",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pendiente",
                "en_preparacion",
                "listo",
                "entregado",
                "cancelado"
              ]
            }
          },
          {
            "name": "id_mesa",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id_usuario",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de pedidos"
          }
        }
      },
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Crear pedido (levantar orden)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_mesa": {
                    "type": "integer",
                    "example": 1
                  },
                  "observaciones": {
                    "type": "string",
                    "example": "Cliente frecuente, mesa junto a la ventana"
                  },
                  "metodo_pago": {
                    "type": "string",
                    "nullable": true
                  },
                  "detalles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id_producto": {
                          "type": "integer",
                          "example": 1
                        },
                        "cantidad": {
                          "type": "integer",
                          "example": 2
                        },
                        "observaciones": {
                          "type": "string",
                          "example": "Sin azucar"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "id_mesa",
                  "detalles"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido registrado"
          },
          "404": {
            "description": "Mesa o producto no encontrado"
          },
          "409": {
            "description": "Producto no disponible"
          }
        }
      }
    },
    "/api/pedidos/{id_pedido}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Obtener detalle del pedido",
        "parameters": [
          {
            "name": "id_pedido",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido con detalles e historial"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Actualizar pedido",
        "description": "Bloqueado si el pedido esta en listo, entregado o cancelado.",
        "parameters": [
          {
            "name": "id_pedido",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "observaciones": {
                    "type": "string",
                    "example": "Agregar un cafe mas"
                  },
                  "metodo_pago": {
                    "type": "string",
                    "example": "tarjeta"
                  },
                  "detalles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id_producto": {
                          "type": "integer"
                        },
                        "cantidad": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido actualizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "409": {
            "description": "Pedido no editable en su estado actual"
          }
        }
      },
      "delete": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Cancelar pedido",
        "description": "Solo si el pedido esta pendiente o en_preparacion.",
        "parameters": [
          {
            "name": "id_pedido",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido cancelado"
          },
          "409": {
            "description": "No se puede cancelar en su estado actual"
          }
        }
      }
    },
    "/api/pedidos/{id_pedido}/estado": {
      "patch": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Cambiar estado del pedido",
        "description": "pendiente -> en_preparacion -> listo (descuenta inventario y notifica al mesero) -> entregado (libera la mesa). cancelado tambien libera la mesa.",
        "parameters": [
          {
            "name": "id_pedido",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": {
                    "type": "string",
                    "enum": [
                      "pendiente",
                      "en_preparacion",
                      "listo",
                      "entregado",
                      "cancelado"
                    ],
                    "example": "listo"
                  },
                  "comentario": {
                    "type": "string",
                    "example": "Terminado a tiempo"
                  }
                },
                "required": [
                  "estado"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "400": {
            "description": "Estado invalido"
          },
          "409": {
            "description": "El pedido ya esta entregado/cancelado"
          }
        }
      }
    },
    "/api/pedidos/{id_pedido}/historial": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Ver historial de estados del pedido",
        "parameters": [
          {
            "name": "id_pedido",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bitacora de cambios de estado"
          }
        }
      }
    },
...
  }
}